pyproject.toml
[project]
name = "voice-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyjokes>=0.8.3",
    "python-dotenv>=1.1.0",
    "pyttsx3>=2.98",
    "requests>=2.32.3",
    "speechrecognition>=3.14.2",
    "wikipedia>=1.4.0",
    "wolframalpha>=5.1.3",
]
